[refactor]: 디렉토리명을 exchanges로 변경
