+++ backend/src/main/java/com/ddbb/dingdong/presentation/endpoint/user/exchanges/UpdateHomeLocationDto.java
@@ -1,4 +1,4 @@
-package com.ddbb.dingdong.presentation.endpoint.user.home;
+package com.ddbb.dingdong.presentation.endpoint.user.exchanges;
 
 import lombok.Getter;
